nodes/dm-text-input/pyproject.toml
[project]
name = "dm-text-input"
version = "0.1.0"
description = "A text input widget for the Interaction UI."
requires-python = ">=3.10"
dependencies = ["dora-rs == 0.4.1", "pyarrow", "requests", "websocket-client"]

[project.scripts]
dm-text-input = "dm_text_input.main:main"
